requirements.txt
beautifulsoup4
bs4
requests
regex
urllib3
sockets
requests_html
googlesearch-python
pybase64
python-math
pycountry
fasttext
numpy
num2words